ecom/requirements_otel.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-django
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp
